requirements.txt
Django>=4.2
djangorestframework
mysqlclient
requests
python-dotenv
whitenoise
gunicorn
